--- a/api/biomarker/backend_utils/db.py
+++ b/api/biomarker/backend_utils/db.py
@@ -217,7 +217,7 @@ def execute_pipeline(
         # TODO : delete logging
         custom_app.api_logger.info(f"PIPELINE:\n{pipeline}")
         explain_output = dbh.command("aggregate", collection, pipeline=pipeline, explain=True)
-        custom_app.api_logger.info(f"COMMAND EXPLAIN OUTPUT:\n{json.dumps(explain_output)}\n\n")
+        custom_app.api_logger.info(f"COMMAND EXPLAIN OUTPUT:\n{explain_output}\n\n")
 
         return result, 200
     except PyMongoError as db_error: